refactor(#633): move AssetBank/TrackBeds + pool/gain consts into scene::audio::bank
The pure sample-bank builders join the synth stack in pixtuoid-scene so the
coming wasm WebAudio driver builds byte-identical banks from the SAME source —
the rng draw-order contract (bank -> rain -> track beds) lives once. AssetBank,
TrackBeds, TRACK_STEMS, and the KEYSTROKE_*/DROP_*/ONE_SHOT_GAIN consts are now
#[doc(hidden)] pub in scene::audio::bank; run_loop imports them. OneShot's
import went test-side (run_loop infers it). Scene 29 + binary 11 audio tests
green verbatim; no behavior change.
